Split off Int256 into separate source.
This should increase readability while maintaining the same API.
